Handle create of blueprint from content to create a new uploaded file. Handle delete of blueprint to delete uploaded files.
