+++ src/Umbraco.Core/Notifications/ContentSavedBlueprintNotification.cs
@@ -5,6 +5,7 @@ using Umbraco.Cms.Core.Events;
 using Umbraco.Cms.Core.Models;
 
 namespace Umbraco.Cms.Core.Notifications;
+
 /// <summary>
 /// A notification that is used to trigger the IContentService when the SavedBlueprint method is called in the API.
 /// </summary>
@@ -14,8 +15,21 @@ public sealed class ContentSavedBlueprintNotification : ObjectNotification<ICont
         : base(target, messages)
     {
     }
+
+    public ContentSavedBlueprintNotification(IContent target, IContent? createdFromContent, EventMessages messages)
+        : base(target, messages)
+    {
+        CreatedFromContent = createdFromContent;
+    }
+
     /// <summary>
     /// Getting the saved blueprint <see cref="IContent"/> object.
     /// </summary>
     public IContent SavedBlueprint => Target;
+
+    /// <summary>
+    /// Getting the saved blueprint <see cref="IContent"/> object.
+    /// </summary>
+    public IContent? CreatedFromContent { get; }
+
 }
