+++ tests/Umbraco.Tests.Integration/Umbraco.Core/Services/TelemetryProviderTests.cs
@@ -116,7 +116,7 @@ internal sealed class TelemetryProviderTests : UmbracoIntegrationTest
         blueprint.SetValue("keywords", "blueprint 3");
         blueprint.SetValue("description", "blueprint 4");
 
-        ContentService.SaveBlueprint(blueprint);
+        ContentService.SaveBlueprint(blueprint, null);
 
         var fromBlueprint = await ContentBlueprintEditingService.GetScaffoldedAsync(blueprint.Key);
         Assert.IsNotNull(fromBlueprint);
